Formatting of code

+++ contracts/TodoList.sol
@@ -1,29 +1,25 @@
 pragma solidity >=0.4.22 <0.8.0;
 
 contract TodoList {
-  uint public taskCount = 0;
+    uint256 public taskCount = 0;
 
-  struct Task {
-      uint id;
-      string content;
-      bool completed;
-  }
+    struct Task {
+        uint256 id;
+        string content;
+        bool completed;
+    }
 
-  mapping(uint => Task) public tasks;
+    mapping(uint256 => Task) public tasks;
 
-  event TaskCreated(
-      uint id,
-      string content,
-      bool completed
-  );
+    event TaskCreated(uint256 id, string content, bool completed);
 
-  constructor() public {
-      createTask("First Task");
-  }
+    constructor() public {
+        createTask("First Task");
+    }
 
-  function createTask(string memory _content) public {
-      taskCount ++;
-      tasks[taskCount] = Task(taskCount, _content, false);
-      emit TaskCreated(taskCount, _content, false);
-  }
+    function createTask(string memory _content) public {
+        taskCount++;
+        tasks[taskCount] = Task(taskCount, _content, false);
+        emit TaskCreated(taskCount, _content, false);
+    }
 }
